refactor: consolidate in-memory-key-manager package into wallet
There's no practical benefit of keeping this separate from the wallet, in the same way
a particular hardware wallet would have, as most wallets will use the in-memory key manager
as the default. It makes more sense to source here in the wallet package.
