--- packages/in-memory-key-manager/.gitignore
@@ -1,2 +1,0 @@
-node_modules
-dist
