--- packages/in-memory-key-manager/README.md
@@ -1,1 +1,0 @@
-# Cardano JS SDK | in-memory-key-manager
