--- packages/in-memory-key-manager/src/.eslintrc.js
@@ -1,7 +1,0 @@
-module.exports = {
-  "extends": ["../../../.eslintrc.js"],
-  "parserOptions": {
-    "project": "./tsconfig.json",
-    "tsconfigRootDir": __dirname
-  }
-}
