--- packages/in-memory-key-manager/src/index.ts
@@ -1,2 +1,0 @@
-export * from './InMemoryKey';
-export * as util from './util';
