--- packages/in-memory-key-manager/src/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "../../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "../dist/",
-    "rootDir": "."
-  },
-  "references": [
-    { "path": "../../core/src" },
-    { "path": "../../wallet/src" }
-  ]
-}
