--- packages/in-memory-key-manager/test/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "extends": "../../../tsconfig.json",
-  "compilerOptions": {
-    "baseUrl": ".",
-    "paths": {
-      "@src/*": ["../src/*"]
-    }
-  },
-  "references": [
-    { "path": "../src" },
-    { "path": "../../core/src" },
-    { "path": "../../wallet/src" }
-  ]
-}
