+++ packages/wallet/test/SingleAddressWallet.test.ts
@@ -3,7 +3,7 @@
 import { createSingleAddressWallet, SingleAddressWallet } from '@src/SingleAddressWallet';
 import * as KeyManagement from '@src/KeyManagement';
 import { loadCardanoSerializationLib, CardanoSerializationLib, Cardano, CardanoProvider } from '@cardano-sdk/core';
-import { createInMemoryKeyManager, util } from '@cardano-sdk/in-memory-key-manager';
+import { createInMemoryKeyManager, util } from '@src/KeyManagement';
 import { InMemoryUtxoRepository } from '@src/InMemoryUtxoRepository';
 import { UtxoRepository } from '@src/UtxoRepository';
 import { InputSelector, roundRobinRandomImprove } from '@cardano-sdk/cip2';
