+++ scripts/pack.sh
@@ -8,5 +8,4 @@ npm pack --cwd ./packages/cip2 && \
 npm pack --cwd ./packages/cip30 && \
 npm pack --cwd ./packages/core && \
 npm pack --cwd ./packages/golden-test-generator && \
-npm pack --cwd ./packages/in-memory-key-manager && \
 npm pack --cwd ./packages/wallet
