+++ scripts/publish.sh
@@ -8,5 +8,4 @@ npm publish --cwd ./packages/cip2 && \
 npm publish --cwd ./packages/cip30 && \
 npm publish --cwd ./packages/core && \
 npm publish --cwd ./packages/golden-test-generator && \
-npm publish --cwd ./packages/in-memory-key-manager && \
 npm publish --cwd ./packages/wallet
